requirements.txt
python-dateutil
python-dotenv
requests 